Updates s3 manifest, fixes bug where search button was hidden on right sidepanel open, fixes auto-scrolling bugs with log viewer
